Split up process_photo into encode + process
