+++ api/scanner/process_photo.go
@@ -3,7 +3,6 @@ package scanner
 import (
 	"database/sql"
 	"fmt"
-	"image"
 	"image/jpeg"
 	"log"
 	"os"
@@ -11,8 +10,7 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/disintegration/imaging"
-	"github.com/h2non/filetype"
+	"github.com/pkg/errors"
 	"github.com/viktorstrate/photoview/api/graphql/models"
 	"github.com/viktorstrate/photoview/api/utils"
 
@@ -23,8 +21,6 @@ import (
 	_ "golang.org/x/image/bmp"
 	_ "golang.org/x/image/tiff"
 	_ "golang.org/x/image/webp"
-
-	cr2Decoder "github.com/nf/cr2"
 )
 
 // Higher order function used to check if PhotoURL for a given PhotoPurpose exists
@@ -52,7 +48,7 @@ func ProcessPhoto(tx *sql.Tx, photo *models.Photo) error {
 
 	log.Printf("Processing photo: %s\n", photo.Path)
 
-	imageData := ProcessImageData{
+	imageData := EncodeImageData{
 		photo: photo,
 	}
 
@@ -61,13 +57,13 @@ func ProcessPhoto(tx *sql.Tx, photo *models.Photo) error {
 	photoBaseName := photoName[0 : len(photoName)-len(path.Ext(photoName))]
 	photoBaseExt := path.Ext(photoName)
 
-	photoChecker, err := makePhotoURLChecker(tx, photo.PhotoID)
+	photoUrlFromDB, err := makePhotoURLChecker(tx, photo.PhotoID)
 	if err != nil {
 		return err
 	}
 
 	// original photo url
-	origURL, err := photoChecker(models.PhotoOriginal)
+	origURL, err := photoUrlFromDB(models.PhotoOriginal)
 	if err != nil {
 		return err
 	}
@@ -96,21 +92,21 @@ func ProcessPhoto(tx *sql.Tx, photo *models.Photo) error {
 	}
 
 	// Thumbnail
-	thumbURL, err := photoChecker(models.PhotoThumbnail)
+	thumbURL, err := photoUrlFromDB(models.PhotoThumbnail)
 	if err != nil {
-		return err
+		return errors.Wrap(err, "error processing thumbnail")
 	}
 
 	// Highres
-	highResURL, err := photoChecker(models.PhotoHighRes)
+	highResURL, err := photoUrlFromDB(models.PhotoHighRes)
 	if err != nil {
-		return err
+		return errors.Wrap(err, "error processing highres")
 	}
 
 	// Make sure photo cache directory exists
 	photoCachePath, err := makePhotoCacheDir(photo)
 	if err != nil {
-		return err
+		return errors.Wrap(err, "cache directory error")
 	}
 
 	// Save thumbnail to cache
@@ -127,8 +123,7 @@ func ProcessPhoto(tx *sql.Tx, photo *models.Photo) error {
 
 		err = encodeImageJPEG(path.Join(*photoCachePath, thumbnail_name), thumbnailImage, &jpeg.Options{Quality: 70})
 		if err != nil {
-			log.Println("ERROR: creating high-res cached image")
-			return err
+			return errors.Wrap(err, "could not create thumbnail cached image")
 		}
 
 		thumbSize := thumbnailImage.Bounds().Max
@@ -136,7 +131,8 @@ func ProcessPhoto(tx *sql.Tx, photo *models.Photo) error {
 		if err != nil {
 			return err
 		}
-	} else if thumbURL != nil {
+	} else {
+		// Verify that thumbnail photo still exists in cache
 		thumbPath := path.Join(*photoCachePath, thumbURL.PhotoName)
 
 		if _, err := os.Stat(thumbPath); os.IsNotExist(err) {
@@ -184,8 +180,7 @@ func ProcessPhoto(tx *sql.Tx, photo *models.Photo) error {
 
 			err = encodeImageJPEG(path.Join(*photoCachePath, highres_name), photoImage, &jpeg.Options{Quality: 70})
 			if err != nil {
-				log.Println("ERROR: creating high-res cached image")
-				return err
+				return errors.Wrap(err, "creating high-res cached image")
 			}
 
 			_, err = tx.Exec("INSERT INTO photo_url (photo_id, photo_name, width, height, purpose, content_type) VALUES (?, ?, ?, ?, ?, ?)",
@@ -195,7 +190,8 @@ func ProcessPhoto(tx *sql.Tx, photo *models.Photo) error {
 				return err
 			}
 		}
-	} else if highResURL != nil {
+	} else {
+		// Verify that highres photo still exists in cache
 		highResPath := path.Join(*photoCachePath, highResURL.PhotoName)
 
 		if _, err := os.Stat(highResPath); os.IsNotExist(err) {
@@ -246,160 +242,4 @@ func makePhotoCacheDir(photo *models.Photo) (*string, error) {
 	}
 
 	return &photoCachePath, nil
-}
-
-func encodeImageJPEG(photoPath string, photoImage image.Image, jpegOptions *jpeg.Options) error {
-	photo_file, err := os.Create(photoPath)
-	if err != nil {
-		log.Printf("ERROR: Could not create file: %s\n", photoPath)
-		return err
-	}
-	defer photo_file.Close()
-
-	err = jpeg.Encode(photo_file, photoImage, jpegOptions)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// ProcessImageData is used to easily decode image data, with a cache so expensive operations are not repeated
-type ProcessImageData struct {
-	photo           *models.Photo
-	_photoImage     image.Image
-	_thumbnailImage image.Image
-	_contentType    *ImageType
-}
-
-// ContentType reads the image to determine its content type
-func (img *ProcessImageData) ContentType() (*ImageType, error) {
-	if img._contentType != nil {
-		return img._contentType, nil
-	}
-
-	file, err := os.Open(img.photo.Path)
-	if err != nil {
-		ScannerError("Could not open file %s: %s\n", img.photo.Path, err)
-		return nil, err
-	}
-	defer file.Close()
-
-	head := make([]byte, 261)
-	if _, err := file.Read(head); err != nil {
-		ScannerError("Could not read photo %s: %s\n", img.photo.Path, err)
-		return nil, err
-	}
-
-	_imgType, err := filetype.Image(head)
-	if err != nil {
-		return nil, err
-	}
-
-	imgType := ImageType(_imgType.MIME.Value)
-	img._contentType = &imgType
-	return img._contentType, nil
-}
-
-// PhotoImage reads and decodes the image file and saves it in a cache so the photo in only decoded once
-func (img *ProcessImageData) PhotoImage(tx *sql.Tx) (image.Image, error) {
-	if img._photoImage != nil {
-		return img._photoImage, nil
-	}
-
-	photoFile, err := os.Open(img.photo.Path)
-	if err != nil {
-		return nil, err
-	}
-	defer photoFile.Close()
-
-	var photoImg image.Image
-	contentType, err := img.ContentType()
-	if err != nil {
-		return nil, err
-	}
-
-	if contentType != nil && *contentType == "image/x-canon-cr2" {
-		photoImg, err = cr2Decoder.Decode(photoFile)
-		if err != nil {
-			return nil, utils.HandleError("cr2 raw image decoding", err)
-		}
-	} else {
-		photoImg, _, err = image.Decode(photoFile)
-		if err != nil {
-			return nil, utils.HandleError("image decoding", err)
-		}
-	}
-
-	// Get orientation from exif data
-	row := tx.QueryRow("SELECT photo_exif.orientation FROM photo JOIN photo_exif WHERE photo.exif_id = photo_exif.exif_id AND photo.photo_id = ?", img.photo.PhotoID)
-	var orientation *int
-	if err = row.Scan(&orientation); err != nil {
-		// If not found use default orientation (not rotate)
-		if err == sql.ErrNoRows {
-			orientation = nil
-		} else {
-			return nil, err
-		}
-	}
-
-	if orientation == nil {
-		defaultOrientation := 0
-		orientation = &defaultOrientation
-	}
-
-	switch *orientation {
-	case 2:
-		photoImg = imaging.FlipH(photoImg)
-		break
-	case 3:
-		photoImg = imaging.Rotate180(photoImg)
-		break
-	case 4:
-		photoImg = imaging.FlipV(photoImg)
-		break
-	case 5:
-		photoImg = imaging.Transpose(photoImg)
-		break
-	case 6:
-		photoImg = imaging.Rotate270(photoImg)
-		break
-	case 7:
-		photoImg = imaging.Transverse(photoImg)
-		break
-	case 8:
-		photoImg = imaging.Rotate90(photoImg)
-		break
-	default:
-		break
-	}
-
-	img._photoImage = photoImg
-	return img._photoImage, nil
-}
-
-// ThumbnailImage downsizes the image and returns it
-func (img *ProcessImageData) ThumbnailImage(tx *sql.Tx) (image.Image, error) {
-	photoImage, err := img.PhotoImage(tx)
-	if err != nil {
-		return nil, err
-	}
-
-	dimensions := photoImage.Bounds().Max
-	aspect := float64(dimensions.X) / float64(dimensions.Y)
-
-	var width, height int
-
-	if aspect > 1 {
-		width = 1024
-		height = int(1024 / aspect)
-	} else {
-		width = int(1024 * aspect)
-		height = 1024
-	}
-
-	thumbImage := imaging.Thumbnail(photoImage, width, height, imaging.NearestNeighbor)
-	img._thumbnailImage = thumbImage
-
-	return img._thumbnailImage, nil
 }
